+++ src/shared/utils/csv.util.ts
@@ -1,0 +1,5 @@
+import { parse } from 'json2csv';
+
+export const toCSV = async (data: any[]) => {
+    return parse(data);
+};