+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [v0.0.1] - 2026-04-08
+- ทำ wind เสร็จแล้ว
+    - GET /wind แสดง wind-data
+        - export to csv
+        - export to csv and delete
+    - GET /wind/history แสดงข้อมูลที่ถูก export เป็น csv
+        - download csv
+        - delete csv
+
 ## [v0.0.0] - 2026-04-03
 
 - `GET /wind-sensor`